refactor: delegate play logic to Controller
